requirements.txt
asgiref==3.7.2
Django==4.2.7
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-currentuser==0.6.1
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-parler==2.3
Pillow==10.1.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
